[c2p_resolver] move OnGCE
