+++ credentials/alts/alts.go
@@ -28,7 +28,6 @@ import (
 	"errors"
 	"fmt"
 	"net"
-	"sync"
 	"time"
 
 	"google.golang.org/grpc/credentials"
@@ -37,6 +36,7 @@ import (
 	"google.golang.org/grpc/credentials/alts/internal/handshaker/service"
 	altspb "google.golang.org/grpc/credentials/alts/internal/proto/grpc_gcp"
 	"google.golang.org/grpc/grpclog"
+	"google.golang.org/grpc/internal/googlecloud"
 )
 
 const (
@@ -54,7 +54,6 @@ const (
 )
 
 var (
-	once          sync.Once
 	maxRPCVersion = &altspb.RpcProtocolVersions_Version{
 		Major: protocolVersionMaxMajor,
 		Minor: protocolVersionMaxMinor,
@@ -148,10 +147,6 @@ func NewServerCreds(opts *ServerOptions) credentials.TransportCredentials {
 }
 
 func newALTS(side core.Side, accounts []string, hsAddress string) credentials.TransportCredentials {
-	once.Do(func() {
-		vmOnGCP = isRunningOnGCP()
-	})
-
 	if hsAddress == "" {
 		hsAddress = hypervisorHandshakerServiceAddress
 	}
@@ -168,7 +163,7 @@ func newALTS(side core.Side, accounts []string, hsAddress string) credentials.Tr
 
 // ClientHandshake implements the client side handshake protocol.
 func (g *altsTC) ClientHandshake(ctx context.Context, addr string, rawConn net.Conn) (_ net.Conn, _ credentials.AuthInfo, err error) {
-	if !vmOnGCP {
+	if !googlecloud.OnGCE() {
 		return nil, nil, ErrUntrustedPlatform
 	}
 
@@ -223,7 +218,7 @@ func (g *altsTC) ClientHandshake(ctx context.Context, addr string, rawConn net.C
 
 // ServerHandshake implements the server side ALTS handshaker.
 func (g *altsTC) ServerHandshake(rawConn net.Conn) (_ net.Conn, _ credentials.AuthInfo, err error) {
-	if !vmOnGCP {
+	if !googlecloud.OnGCE() {
 		return nil, nil, ErrUntrustedPlatform
 	}
 	// Connecting to ALTS handshaker service.
