+++ go.mod
@@ -3,7 +3,6 @@ module google.golang.org/grpc
 go 1.11
 
 require (
-	cloud.google.com/go v0.26.0
 	github.com/cncf/udpa/go v0.0.0-20201120205902-5459f2c99403
 	github.com/envoyproxy/go-control-plane v0.9.9-0.20210217033140-668b12f5399d
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
